+++ cmd/produce/produce_test.go
@@ -3,6 +3,7 @@ package produce_test
 import (
 	"fmt"
 	"github.com/deviceinsight/kafkactl/test_util"
+	"strings"
 	"testing"
 )
 
@@ -146,4 +147,28 @@ func TestProduceFromHexIntegration(t *testing.T) {
 	}
 
 	test_util.AssertEquals(t, "test-key#0000000000000000", kafkaCtl.GetStdOut())
+}
+
+func TestProduceAutoCompletionIntegration(t *testing.T) {
+
+	test_util.StartIntegrationTest(t)
+
+	prefix := "produce-complete-"
+
+	topicName1 := test_util.CreateTopic(t, prefix+"a")
+	topicName2 := test_util.CreateTopic(t, prefix+"b")
+	topicName3 := test_util.CreateTopic(t, prefix+"c")
+
+	kafkaCtl := test_util.CreateKafkaCtlCommand()
+	kafkaCtl.Verbose = false
+
+	if _, err := kafkaCtl.Execute("__complete", "produce", ""); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	outputLines := strings.Split(strings.TrimSpace(kafkaCtl.GetStdOut()), "\n")
+
+	test_util.AssertContains(t, topicName1, outputLines)
+	test_util.AssertContains(t, topicName2, outputLines)
+	test_util.AssertContains(t, topicName3, outputLines)
 }